+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added `sbyte` contracts
+- Added `ushort` contracts
+- Added `ulong` contracts
+- Added `uint` contracts
 
 ## [1.2.0] / 2024-04-29
 - Added a `CAHNGELOG` (this one right here 😏)